requirements.txt
asgiref==3.6.0
Django==4.1.7
Pillow==9.4.0
sqlparse==0.4.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0